fix build for ubuntu-based systems

+++ dep/build.sh
@@ -48,12 +48,14 @@ cd "${builddir}"
 
 dirs="linux asm asm-generic"
 for dir in $dirs; do
-    if [[ -L "/usr/include/$dir" ]]; then
-        # SLES has symlinks for asm
-        realpath=$(readlink -f "/usr/include/$dir")
-        cp -rp "$realpath" "${SPRT_PREFIX}/include/"
+    if [[ -e "/usr/include/$dir" ]]; then
+        if [[ -L "/usr/include/$dir" ]]; then
+            # SLES has symlinks for asm
+            realpath=$(readlink -f "/usr/include/$dir")
+            cp -rp "$realpath" "${SPRT_PREFIX}/include/"
+        fi
+        cp -rp "/usr/include/$dir" "${SPRT_PREFIX}/include/"
     fi
-    cp -rp "/usr/include/$dir" "${SPRT_PREFIX}/include/"
 done
 
 cd "${builddir}"
